requirements.txt
python-dotenv
langchain-openai
langchain-core
streamlit
langchain-community
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
chromadb
faiss-cpu
sentence-transformers
langchain-huggingface
torch
transformers
arxiv
wikipedia
langchainhub
groq
langchain-groq
google-generativeai